# gh-120196: Reuse find_max_char for bytes objects