Simple duplicate code removal and reuse of other function.


<!-- gh-issue-number: gh-120196 -->
* Issue: gh-120196
<!-- /gh-issue-number -->