@vstinner 2024-06-14
It's up to you. This change looks simpler to be merged than the other one which looks controversial.