@serhiy-storchaka 2024-06-14
Currently `bytes.isascii()` returns as soon as it finds the first non-ASCII byte. What with the new implementation?