@vstinner 2024-06-14
> What with the new implementation?

It's about reusing existing code, IMO it's a good thing.