@vstinner 2024-06-17
Merged, thank you. Oops, I merged the PR while @serhiy-storchaka was reviewing it :-)